mnist/requirements.txt
tensorflow>=2.10
keras>=2.10
streamlit>=1.28
pillow>=9.0
numpy>=1.21
opencv-python-headless>=4.5
streamlit-drawable-canvas>=0.9.3
matplotlib>=3.5.1
scipy>=1.16.1
